move IO utils into their own package
