+++ chunky/src/java/se/llbit/chunky/renderer/renderdump/HuffmanDumpFormat.java
@@ -1,7 +1,7 @@
 package se.llbit.chunky.renderer.renderdump;
 
 import se.llbit.chunky.renderer.scene.Scene;
-import se.llbit.util.IsolatedOutputStream;
+import se.llbit.util.io.IsolatedOutputStream;
 
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
