+++ chunky/src/java/se/llbit/util/io/PositionalOutputStream.java
@@ -16,7 +16,7 @@
  * along with Chunky.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package se.llbit.util;
+package se.llbit.util.io;
 
 import se.llbit.util.annotation.NotNull;
